+++ src/features/XIT/ACT/actions/mtra/config.ts
@@ -1,5 +1,3 @@
-export const configuredLocation = 'Configure on Execution';
-
 export interface Config {
   origin?: string;
   destination?: string;